Corrections to scripts for release (Add IF EXISTS etc)

+++ Database_Schema/Mod_Level_Scripts/DDL-224/Tables/c_Database_Status.sql
@@ -1,4 +1,10 @@
-ALTER TABLE c_Database_Status DROP COLUMN vax_component_version
-ALTER TABLE c_Database_Status DROP COLUMN vax_component_id
-ALTER TABLE c_Database_Status ADD client_link varchar(500)
+
+IF (select count(*) from syscolumns where name = 'vax_component_version') > 0
+	ALTER TABLE c_Database_Status DROP COLUMN vax_component_version
+
+IF (select count(*) from syscolumns where name = 'vax_component_id') > 0
+	ALTER TABLE c_Database_Status DROP COLUMN vax_component_id
+
+IF (select count(*) from syscolumns where name = 'client_link') = 0
+	ALTER TABLE c_Database_Status ADD client_link varchar(500)
